Moved distance_computation function and prediction_mapping function from tf_helper_functions.py to hyperparameter_search_MNIST.py. Also fixed a coloring bug in plots.
